Conflicts:
	module/TicketBundle/Hydrator/Event.php

+++ module/TicketBundle/Hydrator/Event.php
@@ -18,9 +18,9 @@
 
 namespace TicketBundle\Hydrator;
 
-use DateTime,
-    TicketBundle\Entity\Event as EventEntity,
-    TicketBundle\Entity\Option;
+use TicketBundle\Entity\Event as EventEntity,
+    TicketBundle\Entity\Option,
+    TicketBundle\Entity\Ticket as TicketEntity;
 
 class Event extends \CommonBundle\Component\Hydrator\Hydrator
 {
@@ -32,7 +32,7 @@ class Event extends \CommonBundle\Component\Hydrator\Hydrator
         'limit_per_person', 'only_members',
     );
 
-    protected function doHydrate(array $array, $object = null)
+    protected function doHydrate(array $data, $object = null)
     {
         if (null === $object) {
             $object = new EventEntity();
@@ -44,10 +44,10 @@ class Event extends \CommonBundle\Component\Hydrator\Hydrator
             ->getRepository('CalendarBundle\Entity\Node\Event')
             ->findOneById($data['event']);
 
-        $closeDate = self::loadDate($data['bookings_close_date']);
+        $closeDate = self::loadDateTime($data['bookings_close_date']);
 
-        $priceMembers = $enableOptions ? 0 : $data['price_members'];
-        $priceNonMembers = $enableOptions ? 0 : ($data['only_members'] ? 0 : $data['price_non_members']);
+        $priceMembers = $enableOptions ? 0 : $data['prices']['price_members'];
+        $priceNonMembers = $enableOptions ? 0 : ($data['only_members'] ? 0 : $data['prices']['price_non_members']);
 
         $generateTickets = $data['generate_tickets'];
 
@@ -81,14 +81,13 @@ class Event extends \CommonBundle\Component\Hydrator\Hydrator
         }
 
         if ($data['generate_tickets']) {
-            if ($event->areTicketsGenerated()) {
+            if ($object->areTicketsGenerated()) {
                 if ($data['number_of_tickets'] >= $object->getNumberOfTickets()) {
                     for ($i = $object->getNumberOfTickets(); $i < $data['number_of_tickets']; $i++) {
                         $this->getEntityManager()->persist(
-                            new Ticket(
+                            new TicketEntity(
                                 $object,
                                 'empty',
-                                null,
                                 null,
                                 null,
                                 null,
@@ -114,10 +113,9 @@ class Event extends \CommonBundle\Component\Hydrator\Hydrator
             } else { // tickets weren't generated yet, but are now
                 for ($i = 0 ; $i < $data['number_of_tickets'] ; $i++) {
                     $this->getEntityManager()->persist(
-                        new Ticket(
+                        new TicketEntity(
                             $object,
                             'empty',
-                            null,
                             null,
                             null,
                             null,
@@ -126,10 +124,10 @@ class Event extends \CommonBundle\Component\Hydrator\Hydrator
                     );
                 }
             }
-        } else { // tickets are not generated (anymore)
+        } elseif ($object->getId()) { // tickets are not generated (anymore)
             $tickets = $this->getEntityManager()
                 ->getRepository('TicketBundle\Entity\Ticket')
-                ->findAllEmptyByEvent($event);
+                ->findAllEmptyByEvent($object);
             foreach ($tickets as $ticket) {
                 $this->getEntityManager()->remove($ticket);
             }
@@ -175,14 +173,5 @@ class Event extends \CommonBundle\Component\Hydrator\Hydrator
         }
 
         return $data;
-    }
-
-    /**
-     * @param  string        $date
-     * @return DateTime|null
-     */
-    protected static function loadDate($date)
-    {
-        return DateTime::createFromFormat('d#m#Y H#i', $date) ?: null;
     }
 }
